requirements.txt
alembic==0.8.5
coverage==4.2
Flask==0.12
Flask-Bcrypt==0.7.1
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
gunicorn==19.4.5
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
nltk==3.2
psycopg2==2.6.1
python-editor==0.5
requests==2.9.1
SQLAlchemy==1.0.12
Werkzeug==0.11.4
autoenv==1.0.0
python_dotenv==0.6.1
wtforms==2.1
gevent==1.1.2
Celery==4.0.2
redis==2.10.5
flower==0.9.1
Flask-Testing==0.6.1
flask-jwt-extended==1.3.1